Revert "fuck unsigned"
This reverts commit 094459093791853c8e5b14ace3e3f15ccf81478a.

+++ plugin_base/src/plugin_base/plugin_base/dsp/splice_engine.cpp
@@ -78,7 +78,7 @@ splice_accurate_events(
   }
 
   // sanity check (this actually did happen)
-  for(int i = 0; i < static_cast<int>(spliced_events.size()) - 1; i++)
+  for(int i = 0; i < spliced_events.size() - 1; i++)
     assert(spliced_events[i] != spliced_events[i + 1]);
 }
